Car detail pages added and bug fixed
Add car detail pages, user can jump into the page by clicking "view details" buttons.

Also fixed a bug:
In the index.html on the homepage, when adding the "View Details" button to each car card, try to generate a jump link in the following way:

<a href="{{ url_for('car_detail', car_id=car['name']) }}">View Details</a>

However, since the car parameter is converted through cars.to _ dict(orient='records') and there is no guarantee that there is a 'name' field (especially when the car _ detail page has not been implemented or tested), the result triggers Jinja's Undefined Error.

Then I changed it to use the line number of the current loop (as car _id), consistent with the car _ detail(car _id) routing logic in Flask:

<a href="{{ url_for('car_detail', car_id=loop.index0) }}" class="btn btn-primary mt-2">View Details</a>

• loop.index0 is the current loop index built in Jinja (starting from 0);

• You can pass in car _id, and get the corresponding vehicle data through cars.iloc[car _id] in the background.